% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/miRLAB.R
\name{experiment}
\alias{experiment}
\title{Function for validate the results from all 12 methods.}
\usage{
experiment(allmethods, topk, Expgroundtruth, LFC, downreg)
}
\arguments{
\item{allmethods}{A list of results (matrix with columns are miRNA and rows are mRNAs).}

\item{topk}{Top k targets of each miRNA that will be extracted for validation}

\item{Expgroundtruth}{The ground truth in .csv file for validation}

\item{LFC}{log fold-change for validating the results using transfection experiments}

\item{downreg}{If set to TRUE the negative effects will have higher ranks than the positives.}
}
\value{
The validation results for all 12 methods
}
\description{
Function for validate the results from all 12 methods.
}

