% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/miRLAB.R
\name{filterAndCompare}
\alias{filterAndCompare}
\title{Filter and compare the validation results from 12 methods
Keep the miRNAs that have at least noVal confirmed targets and compare the validation results from all methods.}
\usage{
filterAndCompare(allresults, noVal)
}
\arguments{
\item{allresults}{the results from all methods generated from experiment function. This is a list.}

\item{noVal}{Number of confirmed targets in each method (threshold) to filter. Records (miRNA) with less than this will be removed}
}
\value{
the validation results of all methods
}
\description{
Filter and compare the validation results from 12 methods
Keep the miRNAs that have at least noVal confirmed targets and compare the validation results from all methods.
}
\examples{
print("result=filterAndCompare(allresults, 2)")
}

