% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRLAB.R
\name{getData}
\alias{getData}
\title{getData from GDC}
\usage{
getData(cancerName)
}
\arguments{
\item{cancerName}{The name of cancer in string format}
}
\value{
dataset in matrix format
}
\description{
getData from GDC
}
\examples{
RawGBM=getData("GBM")
}
