% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/miRLAB.R
\name{GOBPenrichment}
\alias{GOBPenrichment}
\title{Functional enrichment analysis}
\usage{
GOBPenrichment(Genes, Cutoff)
}
\arguments{
\item{Genes}{a list of gene symbols}

\item{Cutoff}{the significant level, e.g. 0.05}
}
\value{
a list of GO terms for the genes
}
\description{
GO BP enrichment analysis for a gene list
}
\examples{
print("result = GOBPenrichment(genelist, 0.05)")
}
\references{
Ashburner, M., Ball, C.A., Blake, J.A., Botstein, D., Butler, H., Cherry, J.M., Davis, A.P., Dolinski, K., Dwight, S.S., Eppig, J.T., Harris, M.A., Hill, D.P., Issel-Tarver, L., Kasarskis, A., Lewis, S., Matese, J.C., Richardson, J.E., Ringwald, M., Rubin, G.M. and Sherlock, G. (2000) Gene Ontology: tool for the unification of biology. Nat. Genet., 25, 25-29.
}

