% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRLAB.R
\name{ICPPam50}
\alias{ICPPam50}
\title{Identify miRNA targets by ICP and PAM50}
\usage{
ICPPam50(d, nmiR, nmR, fiftymRNAsData)
}
\arguments{
\item{d}{A matrix of expression of miRNAs and mRNAs with columns being miRNA or mRNA names and rows being samples}

\item{nmiR}{Number of miRNAs}

\item{nmR}{Number of mRNAs}

\item{fiftymRNAsData}{A matrix of expression of 50 mRNAs in PAM50 with columns being mRNA names and rows being samples}
}
\value{
The matrix of causal effects of miRNAs and mRNAs with columns being miRNAs and rows being mRNAs
}
\description{
This function identifies miRNA targets by ICP and PAM50.
}
\references{
1. Parker, J. S., et al. (2009). "Supervised Risk Predictor of Breast Cancer Based on Intrinsic Subtypes." Journal of Clinical Oncology 27(8): 1160-1167.
}
