% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRLAB.R
\name{identifymiRTargetsByEnsemble}
\alias{identifymiRTargetsByEnsemble}
\title{Identify the top miRNA targets by an ensemble method with ICP-PAM50, Pearson and Lasso}
\usage{
identifymiRTargetsByEnsemble(d, nmiR, nmR, fiftymRNAsData, top = 1,
  topk = 500)
}
\arguments{
\item{d}{A matrix of expression of miRNAs and mRNAs with columns being miRNA or mRNA names and rows being samples}

\item{nmiR}{Number of miRNAs}

\item{nmR}{Number of mRNAs}

\item{fiftymRNAsData}{A matrix of expression of 50 mRNAs in PAM50 with columns being mRNA names and rows being samples}

\item{top}{1 if getting the top of all miRNAs and 2 if getting the top of each miRNA}

\item{topk}{Number of the top to get}
}
\value{
The top k miRNA targets
}
\description{
This function identifies the top miRNA targets by an ensemble method with ICP-PAM50, Pearson and Lasso.
}
\references{
1. Parker, J. S., et al. (2009). "Supervised Risk Predictor of Breast Cancer Based on Intrinsic Subtypes." Journal of Clinical Oncology 27(8): 1160-1167.
}
