% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/miRLAB.R
\name{KEGGenrichment}
\alias{KEGGenrichment}
\title{Functional enrichment analysis
KEGG enrichment analysis for a gene list}
\usage{
KEGGenrichment(Genes, Cutoff)
}
\arguments{
\item{Genes}{a list of gene symbols}

\item{Cutoff}{the significant level, e.g. 0.05}
}
\value{
a list of pathways for the genes
}
\description{
Functional enrichment analysis
KEGG enrichment analysis for a gene list
}
\examples{
print("result = KEGGenrichment(genelist, 0.05)")
}
\references{
Kanehisa, M. and Goto, S. (2000) KEGG: kyoto encyclopedia of genes and genomes. Nucleic Acids Res., 28, 27-30.
}

