\name{miRLAB-package}
\alias{miRLAB-package}
\alias{miRLAB}
\docType{package}
\title{
A dry lab for exploring miRNA-mRNA relationships
}
\description{
Provide tools exploring miRNA-mRNA relationships, including popular miRNA target prediction methods using expression data, ensemble methods that integrate individual methods, functions to get data from online resources, functions to validate the results, and functions to conduct enrichment analyses. 
}
\details{
\tabular{ll}{
Package: \tab miRLAB\cr
Type: \tab Package\cr
Version: \tab 0.99\cr
Date: \tab 2015-04-23\cr
License: \tab GPL(>=2)\cr
}

}
\author{
Thuc Duy Le, Junpeng Zhang

Maintainer: Thuc Duy Le <Thuc.Le@unisa.edu.au>
}
\references{
miRLAB: An R based dry lab for exploring miRNA-mRNA relationships
}
\keyword{
package 
}


