% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/miRLAB.R
\name{Read}
\alias{Read}
\title{Read dataset from csv file}
\usage{
Read(dataset)
}
\arguments{
\item{dataset}{The input dataset in csv format}
}
\value{
dataset in matrix format
}
\description{
Read dataset from csv file
}
\examples{
dataset=system.file("extdata", "ToyEMT.csv", package="miRLAB")
data=Read(dataset)
}

