% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/miRLAB.R
\name{readHeader}
\alias{readHeader}
\title{Read the header of the dataset}
\usage{
readHeader(dataset)
}
\arguments{
\item{dataset}{the character string of the names of the dataset in csv format, e.g. "ToyEMT.csv"}
}
\value{
the header of the dataset
}
\description{
Read the header of the dataset
}
\examples{
dataset=system.file("extdata", "ToyEMT.csv", package="miRLAB")
header=readHeader(dataset)
}

