% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/miRLAB.R
\name{Standardise}
\alias{Standardise}
\title{Stardarsise the dataset
Stadardise the dataset to have mean=0 and std=1 in each column.}
\usage{
Standardise(dataset)
}
\arguments{
\item{dataset}{The input dataset in csv format. e.g. "ToyEMT.csv". The first column is the sample name.}
}
\value{
The standardised dataset.
}
\description{
Stardarsise the dataset
Stadardise the dataset to have mean=0 and std=1 in each column.
}
\examples{
\dontrun{
dataset=system.file("extdata", "ToyEMT.csv", package="miRLAB")
stdata=Standardise(dataset)
}
}

