% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/miRLAB.R
\name{getData}
\alias{getData}
\title{Get data from TCGA}
\usage{
getData(CancerName)
}
\arguments{
\item{CancerName}{The name of cancer using abbreviation of cancer name as one of the followings:
\itemize{
\item Adrenocortical carcinoma: ACC
\item Bladder urothelial carcinoma: BLCA
\item Breast invasive carcinoma: BRCA
\item Cervical squamous cell carcinoma and endocervical adenocarcinoma: CESC
\item Colon adenocarcinoma: COAD
\item Lymphoid neoplasm diffuse large B-cell lymphoma: DLBC
\item Esophageal carcinoma: ESCA
\item Glioblastoma multiforme: GBM
\item Head and neck squamous cell carcinoma: HNSC
\item Kidney chromophobe: KICH
\item Kidney renal clear cell carcinoma: KIRC
\item Kidney renal papillary cell carcinoma: KIRP
\item Acute myeloid leukemia: LAML
\item Brain lower grade glioma: LGG
\item Liver hepatocellular carcinoma: LIHC
\item Lung adenocarcinoma: LUAD
\item Lung squamous cell carcinoma: LUSC
\item Ovarian serous cystadenocarcinoma: OV
\item Pancreatic adenocarcinoma: PAAD
\item Prostate adenocarcinoma: PRAD
\item Rectum adenocarcinoma: READ
\item Sarcoma: SARC
\item Skin cutaneous melanoma: SKCM
\item Stomach adenocarcinoma: STAD
\item Thyroid carcinoma: THCA
\item Uterine corpus endometrial carcinoma: UCEC
\item Uterine carcinosarcoma: UCS
}}
}
\value{
Dataset of matched miRNA and mRNA expression profiles
}
\description{
Download matched miRNA and mRNA expression profiles from TCGA using TCGA Assembler. In order to this function to work, users need to download the
the file "DirectoryTraverseResult_Jul-08-2014.rda" from nugget.unisa.edu.au/Thuc/miRLAB/ and place it in the working directory
}
\examples{
\dontrun{
BreastCancer=getData("BRCA")
}
}
\references{
Zhu, Y., Qiu, P., and Ji, Y. (2014). TCGA-Assembler: open-source software for retrieving and processing TCGA data. Nat. Methods, 11, 599-600.
}

